gsdx: move (most) of the hw hack into a separate file
GSState.cpp was really too big
